requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20230802.0=h313beb8_2
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
arrow-cpp=11.0.0=hc7aafb3_2
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
audioread=3.0.1=py39h2804cbe_1
aws-c-common=0.6.8=h80987f9_1
aws-c-event-stream=0.1.6=h313beb8_6
aws-checksums=0.1.11=h80987f9_2
aws-sdk-cpp=1.8.185=ha71a6ea_1
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=openblas
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.3.5=py39heec5a64_0
braindecode=0.8.1=pypi_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotli-python=1.0.9=py39hc377ac9_7
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=h80987f9_0
ca-certificates=2023.12.12=hca03da5_0
certifi=2024.2.2=py39hca03da5_0
cffi=1.16.0=py39h80987f9_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=py39h48ca7d4_0
cryptography=41.0.7=py39hd4332d6_0
cycler=0.11.0=pyhd3eb1b0_0
datasets=2.18.0=pypi_0
debugpy=1.6.7=py39h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.8=pypi_0
docker-pycreds=0.4.0=pypi_0
docstring-inheritance=2.2.0=pypi_0
einops=0.7.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.13.1=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
gettext=0.21.1=h0186832_0
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gnutls=3.6.15=h887c41c_0
grpc-cpp=1.48.2=hc60591f_4
gtest=1.14.0=h48ca7d4_0
h5py=3.10.0=pypi_0
huggingface-hub=0.21.3=pypi_0
hydra-core=1.3.2=pypi_0
icu=73.1=h313beb8_0
idna=3.4=py39hca03da5_0
importlib-resources=6.1.1=pyhd3eb1b0_1
importlib_resources=6.1.1=py39hca03da5_1
ipykernel=6.28.0=pyh3cd1d5f_0
ipython=8.12.0=pyhd1c38e8_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.2.0=py39hca03da5_0
jpeg=9e=h80987f9_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.5.0=py39hca03da5_0
kiwisolver=1.4.4=py39h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lame=3.100=h1a28f6b_0
lazy_loader=0.3=py39hca03da5_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=8.5.0=h3e2b118_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_0
libflac=1.4.3=hb765f3a_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.17=h0d3ecfb_2
libidn2=2.3.4=h80987f9_0
libllvm14=14.0.6=h7ec7a93_3
libnghttp2=1.57.0=h62f6fdd_0
libogg=1.3.5=h1a28f6b_1
libopenblas=0.3.21=h269037a_0
libopus=1.3.1=h27ca646_1
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
librosa=0.10.1=pyhd8ed1ab_0
libsndfile=1.2.2=h9739721_1
libsodium=1.0.18=h27ca646_1
libsqlite=3.44.2=h091b4b1_0
libssh2=1.10.0=h02f6b3c_2
libtasn1=4.19.0=h80987f9_0
libthrift=0.15.0=h73c2103_2
libtiff=4.5.1=h313beb8_0
libunistring=0.9.10=h1a28f6b_0
libvorbis=1.3.7=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=14.0.6=hc6e5704_0
llvmlite=0.41.0=py39h514c7bf_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.2=py39hdf13c20_0
matplotlib-base=3.8.2=py39h1a09f3e_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mne=1.6.1=pypi_0
mpg123=1.32.4=hebf3989_0
mpmath=1.3.0=pypi_0
msgpack-python=1.0.3=py39h525c30c_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=h84b5d62_1
networkx=3.2.1=pypi_0
numba=0.58.1=py39h313beb8_0
numexpr=2.8.7=py39hecc3335_0
numpy=1.26.3=py39h3b2db8e_0
numpy-base=1.26.3=py39ha9811e2_0
omegaconf=2.3.0=pypi_0
openh264=1.8.0=h98b2900_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.2.1=h0d3ecfb_0
orc=1.7.4=hdca1487_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=py39h46d7db6_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.1=py39h3b245a6_0
pip=23.3.1=py39hca03da5_0
platformdirs=4.1.0=pyhd8ed1ab_0
pooch=1.7.0=py39hca03da5_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py39h1a28f6b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=15.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pyhd8ed1ab_0
pyopenssl=23.2.0=py39hca03da5_0
pyparsing=3.0.9=py39hca03da5_0
pysocks=1.7.1=py39hca03da5_0
pysoundfile=0.12.1=pyhd8ed1ab_0
python=3.9.18=hfa1ae8a_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2023.3.post1=py39hca03da5_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py39h313beb8_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
regex=2023.12.25=pypi_0
requests=2.31.0=py39hca03da5_0
safetensors=0.4.2=pypi_0
scikit-learn=1.3.0=py39h46d7db6_1
scipy=1.11.4=py39h20cbe94_0
sentry-sdk=1.40.6=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py39hca03da5_0
six=1.16.0=pyhd3eb1b0_1
skorch=0.15.0=pypi_0
smmap=5.0.1=pypi_0
snappy=1.1.10=h313beb8_1
soxr=0.1.3=h5008568_3
soxr-python=0.3.7=py39h373d45f_0
sqlite=3.41.2=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h48ca7d4_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.15.2=pypi_0
torch=2.2.1=pypi_0
torcheval=0.0.7=pypi_0
torchinfo=1.8.0=pypi_0
tornado=6.1=py39h5161555_1
tqdm=4.66.2=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
transformers=4.38.2=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023d=h04d1e81_0
urllib3=1.26.18=py39hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
wandb=0.16.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.41.2=py39hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xxhash=3.4.1=pypi_0
xz=5.4.5=h80987f9_0
yarl=1.9.4=pypi_0
zeromq=4.3.4=hc377ac9_0
zipp=3.17.0=py39hca03da5_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=hd90d995_0
